+++ src/Http/Requests/Admin/UpdateBidRequest.php
@@ -3,12 +3,14 @@
 namespace Modules\Custom\MakerBid\Http\Requests\Admin;
 
 use Illuminate\Foundation\Http\FormRequest;
+use Modules\Custom\MakerBid\Http\Concerns\FlattensValidationErrors;
 use Modules\Custom\MakerBid\Support\BidRules;
 use Modules\Custom\MakerBid\Support\BlankToNull;
 
 class UpdateBidRequest extends FormRequest
 {
     use BlankToNull;
+    use FlattensValidationErrors;
 
     public function authorize(): bool
     {
@@ -17,6 +19,7 @@ class UpdateBidRequest extends FormRequest
 
     protected function prepareForValidation(): void
     {
+        $this->liftNestedFormFields(['form', 'edit']);
         $this->nullBlankFields(['days', 'message']);
     }